Remove the `checkScalaJSSemantics` feature.
The implementation of that feature was piggybacked on the
`JS_DEPENDENCIES` file maintained for the `jsDependencies`
mechanism. As the latter will be extracted in a separate sbt
plugin, the required semantics information will not be part of
all Scala.js artifacts, and will therefore be bogus. This means
that the `checkScalaJSSemantics` feature would not do its job
correctly anyway, and it is better to remove it than to give a
false sense of security.

An alternative would be to store the relevant information in a
separate file, but I do not believe it is worth it. Reasonable
Scala.js libraries should never require compliant semantics
anyway. Moreover, I have never had any feedback mentioning that
feature, which probably means no one ever saw that feature in
action, so it is useless.